requirements.txt
chromadb
datasets
langchain-chroma
langchain-core
langchain-openai
python-dotenv
ragas
